Refactor: migrate coherence updates into active modules
Intent: streamline coherence metric tracking
Operators involved: [Coherence, Self-organization]
Affected invariants: [#4, #7]
Key changes:
- move _update_coherence and related helpers into metrics/coherence
- drop obsolete metrics/coherence_updates.py and redirect imports
- update metrics/__init__ and core to use new locations
Expected risks/dissonances: possible metric omissions contained by full test suite
Metrics: C(t), Si, νf, phase expected unchanged
Equivalence map: _update_* functions now exported from metrics.coherence
